+++ src/Cabin/Bridge/Filter/Blog/DeleteCategoryFilter.php
@@ -1,0 +1,27 @@
+<?php
+declare(strict_types=1);
+namespace Airship\Cabin\Bridge\Filter\Blog;
+
+use Airship\Engine\Security\Filter\{
+    BoolFilter,
+    InputFilterContainer,
+    IntFilter
+};
+
+/**
+ * Class DeleteCategoryFilter
+ * @package Airship\Cabin\Bridge\Filter\Account
+ */
+class DeleteCategoryFilter extends InputFilterContainer
+{
+    /**
+     * DeletePostFilter constructor.
+     */
+    public function __construct()
+    {
+        $this
+            ->addFilter('confirm', new BoolFilter())
+            ->addFilter('moveChildrenTo', new IntFilter())
+        ;
+    }
+}